requirements.txt
aiohttp==3.1.3
async-timeout==2.0.1
attrs==17.4.0
chardet==3.0.4
click==3.3
greenlet==0.4.5
idna==2.6
idna-ssl==1.0.1
multidict==4.2.0
Pillow==5.1.0
redis==2.10.3
rq==0.5.0
yarl==1.1.1
